# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VPN Subscription Merger API
  description: |
    Comprehensive API for VPN subscription merging, configuration generation, and management.
    
    ## Features
    - Multi-source VPN subscription aggregation
    - Configuration generation for multiple protocols
    - Real-time monitoring and analytics
    - GraphQL API for advanced queries
    - Web-based configuration generator
    
    ## Authentication
    Most endpoints require no authentication. Some administrative endpoints may require API tokens.
    
    ## Rate Limiting
    API calls are rate-limited to prevent abuse. Default limits:
    - 100 requests per minute for most endpoints
    - 10 requests per minute for resource-intensive operations
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: VPN Merger Support
    url: https://github.com/your-org/CleanConfigs-SubMerger
  license:
    name: GPL-3.0-or-later
    url: https://spdx.org/licenses/GPL-3.0-or-later.html
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.vpnmerger.com
    description: Production server
paths:
...
  /api/nodes.json:
    get:
      summary: Get VPN Nodes
      description: Retrieve available VPN nodes with filtering and sorting
      tags:
        - Free Nodes Aggregator
      parameters:
        - name: limit
          in: query
          description: Maximum number of nodes to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: sort
          in: query
          description: Sort order for nodes
          required: false
          schema:
            type: string
            enum: [score, latency, country, protocol]
            default: score
        - name: protocol
          in: query
          description: Filter by protocol
          required: false
          schema:
            type: string
            enum: [vless, vmess, trojan, shadowsocks]
        - name: country
          in: query
          description: Filter by country code
          required: false
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
      responses:
        '200':
          description: Nodes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodesResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/sources:
    post:
      summary: Add Source URLs
      description: Add new source URLs to the aggregator
      tags:
        - Free Nodes Aggregator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcesRequest'
      responses:
        '200':
          description: Sources added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcesResponse'
        '400':
          description: Invalid source URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...
    Node:
      type: object
      properties:
        id:
          type: string
          example: "node_123"
        host:
          type: string
          example: "example.com"
        port:
          type: integer
          example: 443
        protocol:
          type: string
          enum: [vless, vmess, trojan, shadowsocks]
          example: "vless"
        country:
          type: string
          example: "US"
        score:
          type: number
          format: float
          example: 0.85
        latency:
          type: number
          format: float
          example: 45.2
        last_checked:
          type: string
          format: date-time
        config:
          type: string
          description: Full configuration string
    NodesResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        total:
          type: integer
          example: 150
        filtered:
          type: integer
          example: 100
    SourcesRequest:
      type: object
      required:
        - urls
      properties:
        urls:
          type: array
          items:
            type: string
            format: uri
          example: ["https://example.com/free.txt", "https://another.com/nodes.txt"]
    SourcesResponse:
      type: object
      properties:
        added:
          type: integer
          example: 2
        skipped:
          type: integer
          example: 0
        errors:
          type: array
          items:
            type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if required)
...